support describe table command with only column name and data types
