--- core/src/main/scala/org/apache/spark/sql/execution/command/TiShowTablesCommand.scala
@@ -1,26 +1,0 @@
-package org.apache.spark.sql.execution.command
-
-import org.apache.spark.sql.{Row, SparkSession, TiContext}
-
-case class TiShowTablesCommand(tiContext: TiContext, delegate: ShowTablesCommand)
-    extends TiCommand(delegate) {
-  override def run(sparkSession: SparkSession): Seq[Row] = {
-    val db = delegate.databaseName.getOrElse(tiCatalog.getCurrentDatabase)
-    // Show the information of tables.
-    val tables =
-      delegate.tableIdentifierPattern
-        .map(tiCatalog.listTables(db, _))
-        .getOrElse(tiCatalog.listTables(db))
-    tables.map { tableIdent =>
-      val database = tableIdent.database.getOrElse("")
-      val tableName = tableIdent.table
-      val isTemp = tiCatalog.isTemporaryTable(tableIdent)
-      if (delegate.isExtended) {
-        val information = tiCatalog.getTempViewOrPermanentTableMetadata(tableIdent).simpleString
-        Row(database, tableName, isTemp, s"$information\n")
-      } else {
-        Row(database, tableName, isTemp)
-      }
-    }
-  }
-}
