+++ core/src/main/scala/org/apache/spark/sql/extensions/rules.scala
@@ -1,3 +1,17 @@
+/*
+ * Copyright 2018 PingCAP, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.apache.spark.sql.extensions
 
 import com.pingcap.tispark.statistics.StatisticsManager
@@ -8,6 +22,7 @@ import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.catalyst.rules.Rule
 import com.pingcap.tispark.{MetaManager, TiDBRelation, TiTableReference}
 import org.apache.spark.sql.catalyst.catalog.TiSessionCatalog
+import org.apache.spark.sql.execution.command._
 import org.apache.spark.sql.execution.datasources.LogicalRelation
 
 case class TiResolutionRule(getOrCreateTiContext: SparkSession => TiContext)(
@@ -50,5 +65,22 @@ case class TiResolutionRule(getOrCreateTiContext: SparkSession => TiContext)(
       } else {
         rel
       }
+  }
+}
+
+case class TiDDLRule(getOrCreateTiContext: SparkSession => TiContext)(sparkSession: SparkSession)
+    extends Rule[LogicalPlan] {
+  protected lazy val tiContext: TiContext = getOrCreateTiContext(sparkSession)
+
+  override def apply(plan: LogicalPlan): LogicalPlan = plan transformUp {
+    // TODO: support other commands that may concern TiSpark catalog.
+    case sd: ShowDatabasesCommand =>
+      TiShowDatabasesCommand(tiContext, sd)
+    case sd: SetDatabaseCommand =>
+      TiSetDatabaseCommand(tiContext, sd)
+    case st: ShowTablesCommand =>
+      TiShowTablesCommand(tiContext, st)
+    case dt: DescribeTableCommand =>
+      TiDescribeTablesCommand(tiContext, dt)
   }
 }
